Stopped using np.int and np.float. Modified ravel wrapper to work with gym weird inf casting behavior

+++ abmarl/sim/corridor/multi_corridor.py
@@ -30,7 +30,7 @@ class MultiCorridor(AgentBasedSimulation):
                 id=f'agent{i}',
                 action_space=Discrete(3),
                 observation_space={
-                    'position': Box(0, self.end-1, (1,), np.int),
+                    'position': Box(0, self.end-1, (1,), int),
                     'left': MultiBinary(1),
                     'right': MultiBinary(1)
                 }
